chore(shared_memory): moved under interpreter crate; small refactoring; doc comments
