+++ crates/interpreter/src/instructions/macros.rs
@@ -49,46 +49,15 @@ macro_rules! gas_or_fail {
     };
 }
 
-// macro_rules! memory_resize {
-//     ($interp:expr, $offset:expr, $len:expr) => {{
-//         let len: usize = $len;
-//         let offset: usize = $offset;
-//         if let Some(new_size) =
-//             crate::interpreter::memory::next_multiple_of_32(offset.saturating_add(len))
-//         {
-//             #[cfg(feature = "memory_limit")]
-//             if new_size > ($interp.memory_limit as usize) {
-//                 $interp.instruction_result = InstructionResult::MemoryLimitOOG;
-//                 return;
-//             }
-//
-//             if new_size > $interp.memory.len() {
-//                 if crate::USE_GAS {
-//                     let num_bytes = new_size / 32;
-//                     if !$interp.gas.record_memory(crate::gas::memory_gas(num_bytes)) {
-//                         $interp.instruction_result = InstructionResult::MemoryLimitOOG;
-//                         return;
-//                     }
-//                 }
-//                 $interp.memory.resize(new_size);
-//             }
-//         } else {
-//             $interp.instruction_result = InstructionResult::MemoryOOG;
-//             return;
-//         }
-//     }};
-// }
-
 macro_rules! shared_memory_resize {
     ($interp:expr, $offset:expr, $len:expr) => {{
         let len: usize = $len;
         let offset: usize = $offset;
         let mut mem = $interp.shared_memory.borrow_mut();
         if let Some(new_size) =
-            crate::interpreter::memory::next_multiple_of_32(offset.saturating_add(len))
+            crate::interpreter::shared_memory::next_multiple_of_32(offset.saturating_add(len))
         {
-            #[cfg(feature = "memory_limit")]
-            if new_size > (mem.limit as usize) {
+            if new_size as u64 > mem.limit {
                 $interp.instruction_result = InstructionResult::MemoryLimitOOG;
                 return;
             }
