+++ crates/revm/benches/bench.rs
@@ -7,7 +7,7 @@ use revm::{
     interpreter::{analysis::to_analysed, BytecodeLocked, Contract, DummyHost, Interpreter},
     primitives::{BerlinSpec, Bytecode, BytecodeState, TransactTo, U256},
 };
-use revm_interpreter::primitives::shared_memory::SharedMemory;
+use revm_interpreter::SharedMemory;
 use std::{cell::RefCell, rc::Rc, time::Duration};
 
 type Evm = revm::EVM<BenchmarkDB>;
@@ -104,6 +104,7 @@ fn bench_transact(g: &mut BenchmarkGroup<'_, WallTime>, evm: &mut Evm) {
 }
 
 fn bench_eval(g: &mut BenchmarkGroup<'_, WallTime>, evm: &Evm) {
+    let shared_memory = Rc::new(RefCell::new(SharedMemory::new(evm.env.tx.gas_limit, None)));
     g.bench_function("eval", |b| {
         let contract = Contract {
             input: evm.env.tx.data.clone(),
@@ -111,7 +112,6 @@ fn bench_eval(g: &mut BenchmarkGroup<'_, WallTime>, evm: &Evm) {
             ..Default::default()
         };
         let mut host = DummyHost::new(evm.env.clone());
-        let shared_memory = Rc::new(RefCell::new(SharedMemory::new(evm.env.tx.gas_limit, None)));
         b.iter(|| {
             let mut interpreter =
                 Interpreter::new(Box::new(contract.clone()), u64::MAX, false, &shared_memory);
